pyproject.toml
[tool.poetry]
name = "MaStR"
version = "1.1.0"
description = "Module to access the API of Marktstammdatenregister."
authors = ["Alexander Blinn"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pandas = "^2.2.3"
zeep = "^4.3.1"
matplotlib = "^3.10.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
ruff = "^0.8.3"
ipykernel = "^6.29.5"

[tool.pytest.ini_options]
pythonpath = "app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"